requirements.txt
Pillow~=9.5.0
opencv-python~=4.9.0.80
numpy~=1.24.3
ultralytics~=8.0.232
setuptools~=60.2.0